NMS-10533 rename package topogen => generator
