+++ features/enlinkd/org.opennms.features.enlinkd.topogen/src/main/java/org/opennms/enlinkd/generator/TopologyPersister.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.enlinkd.topogen;
+package org.opennms.enlinkd.generator;
 
 import java.sql.SQLException;
 import java.util.Arrays;
