+++ features/enlinkd/org.opennms.features.enlinkd.topogen/src/main/java/org/opennms/enlinkd/generator/protocol/LldpProtocol.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.enlinkd.topogen.protocol;
+package org.opennms.enlinkd.generator.protocol;
 
 import java.sql.SQLException;
 import java.util.ArrayList;
@@ -36,9 +36,9 @@ import java.util.UUID;
 
 import org.apache.commons.lang3.tuple.Pair;
 import org.opennms.core.utils.LldpUtils;
-import org.opennms.enlinkd.topogen.TopologyContext;
-import org.opennms.enlinkd.topogen.TopologyGenerator;
-import org.opennms.enlinkd.topogen.topology.PairGenerator;
+import org.opennms.enlinkd.generator.TopologyContext;
+import org.opennms.enlinkd.generator.TopologyGenerator;
+import org.opennms.enlinkd.generator.topology.PairGenerator;
 import org.opennms.netmgt.enlinkd.model.LldpElement;
 import org.opennms.netmgt.enlinkd.model.LldpLink;
 import org.opennms.netmgt.model.OnmsNode;
