+++ features/enlinkd/org.opennms.features.enlinkd.topogen/src/main/java/org/opennms/enlinkd/generator/util/RandomUtil.java
@@ -1,8 +1,8 @@
 /*******************************************************************************
  * This file is part of OpenNMS(R).
  *
- * Copyright (C) 2019 The OpenNMS Group, Inc.
- * OpenNMS(R) is Copyright (C) 1999-2019 The OpenNMS Group, Inc.
+ * Copyright (C) 2018 The OpenNMS Group, Inc.
+ * OpenNMS(R) is Copyright (C) 1999-2018 The OpenNMS Group, Inc.
  *
  * OpenNMS(R) is a registered trademark of The OpenNMS Group, Inc.
  *
@@ -26,44 +26,16 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.enlinkd.topogen.topology;
+package org.opennms.enlinkd.generator.util;
 
 import java.util.List;
+import java.util.Random;
 
-import org.apache.commons.lang3.tuple.Pair;
+public class RandomUtil {
 
-public class LinkedPairGenerator<E> implements PairGenerator<E> {
+    private final Random random = new Random(42);
 
-    private final List<E> elements;
-    private final int lastIndexInList;
-    private int indexLeft = -1;
-    private int indexRight = 0;
-
-    @Override
-    public Pair<E, E> next() {
-        compute();
-        return Pair.of(elements.get(indexLeft), elements.get(indexRight));
+    public <E> E getRandom(List<E> collection){
+        return collection.get(random.nextInt(collection.size()));
     }
-
-    public LinkedPairGenerator(List<E> elements) {
-        if (elements == null || elements.size() < 2) {
-            throw new IllegalArgumentException("Need at least 2 elements in list to make a pair");
-        }
-        this.elements = elements;
-        lastIndexInList = elements.size() - 1;
-
-    }
-
-    private void compute() {
-        indexLeft = next(indexLeft);
-        indexRight = next(indexRight);
-    }
-
-    private int next(int i) {
-        if (i == lastIndexInList) {
-            return 0;
-        }
-        return ++i;
-    }
-
 }
